social/requirements.txt
Django==2.2.4
django-phone-field==1.8.0
django-rest-registration==0.5.3
djangorestframework==3.10.3
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
typing==3.7.4.1
